Move util functions from hooks to utils file
