+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_TripSearchScreen/use-trips-query.ts
@@ -15,10 +15,7 @@ import {DateString, SearchTime} from '@atb/journey-date-picker';
 import {TripSearchPreferences, usePreferences} from '@atb/preferences';
 import {useRemoteConfig} from '@atb/RemoteConfigContext';
 import {useSearchHistory} from '@atb/search-history';
-import {
-  SearchStateType,
-  TripPatternWithKey,
-} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/types';
+import {SearchStateType} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/types';
 import {flatMap} from '@atb/utils/array';
 import {enumFromString} from '@atb/utils/enum-from-string';
 import {isValidTripLocations} from '@atb/utils/location';
@@ -28,6 +25,7 @@ import {useCallback, useEffect, useRef, useState} from 'react';
 import {useJourneyModes} from './hooks';
 import {useAnalytics} from '@atb/analytics';
 import {TravelSearchFiltersSelectionType} from '@atb/travel-search-filters';
+import {TripPatternWithKey} from '@atb/travel-details-screens/types';
 
 export function useTripsQuery(
   fromLocation: Location | undefined,
