+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/index.tsx
@@ -1,3 +1,2 @@
 export {TabNav_DashboardStack} from './TabNav_DashboardStack';
 export type {SearchForLocations} from './types';
-export {getBookingRequirementForLeg} from './Dashboard_TripSearchScreen/hooks';
